requirements.txt
flask
twilio
azure-cognitiveservices-speech
openai